Adding model file withing charityprojects
