+++ backend/src/charityprojects/admin.py
@@ -1,5 +1,7 @@
 from django.contrib import admin
-from .models import CharityProjects
+from .models import CharityProjects, Prize
+
 # Register your models here.
 
 admin.site.register(CharityProjects)
+admin.site.register(Prize)
