+++ backend/src/charityprojects/models.py
@@ -47,6 +47,14 @@ class ProjectUser(models.Model):
         return user_email
 
 
+class Prize(models.Model):
+    Category = models.CharField(max_length=120)
+    Tags = models.CharField(max_length=120)
+    ImageName = models.CharField(max_length=120)
+    Image = models.ImageField(upload_to='prizeImages')
+
+    def __str__(self):
+        return '{} {} {} {}'.format(self.Category,  self.Tags, self.ImageName, self.Image)
