--- backend/src/prize/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
-from .models import Prize
-admin.site.register(Prize)
