--- backend/src/prize/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class PrizeConfig(AppConfig):
-    name = 'prize'
