+++ backend/src/ucc/settings.py
@@ -40,7 +40,6 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'myaccount',
-    'prize',
     'corsheaders',
     'rest_framework',
     'accounts',
